requirements.txt
websockets==11.0.3
pyyaml==6.0
prometheus_client==0.16.0
